Don't spew 'Invalid gralloc handle...'.
The message is spewed from unlockPreviousOverlayBuffer in hwcomposer.cpp.

hwcomposer just holds pointer to private_handle_t and can not detect
the pointer has been already deleted, so if someone deletes the
private_handle_t, the message will always spew in log.

To prevent this message, the pointer is judged 'deleted' when
magic equals '0', and the message isn't spewed.